Tests: Initial PR to move important check to package
