+++ lib/spack/spack/cmd/test.py
@@ -7,7 +7,6 @@ from __future__ import print_function
 
 import argparse
 import fnmatch
-import inspect
 import os
 import re
 import shutil
@@ -207,24 +206,13 @@ environment variables:
                    fail_first=args.fail_first)
 
 
-def has_test_method(pkg):
-    if not inspect.isclass(pkg):
-        tty.die('{0}: is not a class, it is {1}'.format(pkg, type(pkg)))
-
-    pkg_base = spack.package.PackageBase
-    return (
-        (issubclass(pkg, pkg_base) and pkg.test != pkg_base.test) or
-        (isinstance(pkg, pkg_base) and pkg.test.__func__ != pkg_base.test)
-    )
-
-
 def test_list(args):
     """List installed packages with available tests."""
     tagged = set(spack.repo.path.packages_with_tags(*args.tag)) if args.tag \
         else set()
 
     def has_test_and_tags(pkg_class):
-        return has_test_method(pkg_class) and \
+        return spack.package.has_test_method(pkg_class) and \
             (not args.tag or pkg_class.name in tagged)
 
     if args.list_all:
